Change is_reachable and is_observable
The observability and reachability is determined by the rank of the observability and reachability matrices. The new implementation uses the np.linalg.matrix_rank function that is based on the svd. This also allows for the use of a tolerance.
The default tolerance is based on the fact that the rank of the Hankel matrix is the relevant property. If we suppose that the system is a balanced realizations (or at least close to it, as we would like) the tolerance is related to the tolerance for the rank of the Hankel matrix by $\sigma_H=\sigma_O\sigma_R$.
Accordingly we set $tol_O = tol_R \sqrt{tol_H}$