Simplified the implementation of reduction
The new implementation only uses the matrices of one stage. This makes it faster as it does not require a SCD of O and R. On the other hand the meaning of the tolerance is not well defined. This makes it harder to choose an apropriate tolerance. But for cases where only a reduction of the computatioal cost is required the function works very well.